pyproject.toml
[project]
name = "personal-ai-employee-hackathon-0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-api-python-client>=2.194.0",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
    "mcp>=1.27.0",
    "odoorpc>=0.10.0",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "watchdog>=6.0.0",
]
